Position double

+++ src/engine/geom/Position.java
@@ -7,21 +7,21 @@ public class Position {
     /**
      * The x coordinate of the point.
      */
-    public int x;
+    public double x;
     
     /**
      * The y coordinate of the point.
      */
-    public int y;
+    public double y;
 
     /**
      * Constructs and initializes a {@code Position} with the specified coordinates.
      *
      * @param x the specified x coordinate
      * @param y the specified y coordinate
-     * @see setLocation(int x, int y)
+     * @see setLocation(double x, double y)
      */
-    public Position(int x, int y) {
+    public Position(double x, double y) {
         setLocation(x, y);
     }
 
@@ -29,9 +29,9 @@ public class Position {
      * Returns the x coordinate of the point.
      * 
      * @return the x coordinate
-     * @see setX(int x)
+     * @see setX(double x)
      */
-    public int getX() {
+    public double getX() {
         return x;
     }
 
@@ -39,9 +39,9 @@ public class Position {
      * Returns the y coordinate of the point.
      * 
      * @return the y coordinate
-     * @see setY(int y)
+     * @see setY(double y)
      */
-    public int getY() {
+    public double getY() {
         return y;
     }
 
@@ -51,7 +51,7 @@ public class Position {
      * @param x the new x coordinate
      * @see getX()
      */
-    public void setX(int x) {
+    public void setX(double x) {
         this.x = x;
     }
 
@@ -61,7 +61,7 @@ public class Position {
      * @param y the new x coordinate
      * @see getY()
      */
-    public void setY(int y) {
+    public void setY(double y) {
         this.y = y;
     }
 
@@ -70,10 +70,10 @@ public class Position {
      *
      * @param x the new x coordinate
      * @param y the new y coordinate
-     * @see setX(int x)
-     * @see setY(int y)
+     * @see setX(double x)
+     * @see setY(double y)
      */
-    public void setLocation(int x, int y) {
+    public void setLocation(double x, double y) {
         setX(x);
         setY(y);
     }
@@ -82,7 +82,7 @@ public class Position {
      * Gets the location of the point.
      *
      * @return a new instance of {@code Position} with the same coordinates.
-     * @see setLocation(int x, int y)
+     * @see setLocation(double x, double y)
      */
     public Position getLocation() {
         return new Position(x, y);
@@ -93,10 +93,10 @@ public class Position {
      *
      * @param dx the x coordinate to add
      * @param dy the y coordinate to add
-     * @see translateX(int dx)
-     * @see translateY(int dy)
+     * @see translateX(double dx)
+     * @see translateY(double dy)
      */
-    public void translate(int dx, int dy) {
+    public void translate(double dx, double dy) {
         translateX(dx);
         translateY(dy);
     }
@@ -105,9 +105,9 @@ public class Position {
      * Translates this point, at location (x,y) by {@code dx} along the x axis
      * 
      * @param dx the x coordinate to add
-     * @see setX(int x)
+     * @see setX(double x)
      */
-    public void translateX(int dx) {
+    public void translateX(double dx) {
         setX(x + dx);
     }
 
@@ -115,9 +115,9 @@ public class Position {
      * Translates this point, at location (x,y) by {@code dy} along the y axis
      * 
      * @param dy the y coordinate to add
-     * @see setY(int y)
+     * @see setY(double y)
      */
-    public void translateY(int dy) {
+    public void translateY(double dy) {
         setY(y + dy);
     }
